check the cookie to see if the user is logged in when the page loads, and persist that cookie into redux state so other components can get at it
